+++ src/translations/nbTranslations.json
@@ -1861,6 +1861,10 @@
           "Vinyl": "LP/EP"
         },
         "music_score_ensemble": "Besetning",
+        "music_score_format": {
+          "minutes": "MM:SS",
+          "hours": "TT:MM:SS"
+        },
         "music_score_ismn": "ISMN",
         "music_score_ismn_helper_text": "Består av 12 tall. Eksempel: 9790260000438",
         "music_score_isrc": "ISRC",
